[fix] updated hero on homepage
